Rename frontend page Project to Application
