+++ frontend/src/components/DevicesBrowser.vue
@@ -151,14 +151,13 @@ import { PlusSmIcon } from '@heroicons/vue/solid'
 
 import { markRaw } from 'vue'
 
+import ProjectStatusBadge from '../pages/application/components/ProjectStatusBadge'
+import ApplicationLink from '../pages/application/components/cells/ApplicationLink'
+import DeviceLink from '../pages/application/components/cells/DeviceLink.vue'
+import InstanceInstancesLink from '../pages/application/components/cells/InstanceInstancesLink.vue'
+import Snapshot from '../pages/application/components/cells/Snapshot.vue'
 import DeviceLastSeenBadge from '../pages/device/components/DeviceLastSeenBadge'
 import SnapshotAssignDialog from '../pages/instance/Snapshots/dialogs/SnapshotAssignDialog'
-
-import ProjectStatusBadge from '../pages/project/components/ProjectStatusBadge'
-import ApplicationLink from '../pages/project/components/cells/ApplicationLink'
-import DeviceLink from '../pages/project/components/cells/DeviceLink.vue'
-import InstanceInstancesLink from '../pages/project/components/cells/InstanceInstancesLink.vue'
-import Snapshot from '../pages/project/components/cells/Snapshot.vue'
 import DeviceAssignInstanceDialog from '../pages/team/Devices/dialogs/DeviceAssignInstanceDialog'
 import DeviceCredentialsDialog from '../pages/team/Devices/dialogs/DeviceCredentialsDialog'
 import TeamDeviceCreateDialog from '../pages/team/Devices/dialogs/TeamDeviceCreateDialog'
