+++ frontend/src/pages/application/Logs.vue
