+++ frontend/src/pages/application/Settings/dialogs/ConfirmApplicationDeleteDialog.vue
