+++ frontend/src/pages/application/Snapshots/components/cells/DeviceCount.vue
