+++ frontend/src/pages/application/Snapshots/components/cells/SnapshotName.vue
