+++ frontend/src/pages/application/components/cells/ApplicationLink.vue
