+++ frontend/src/pages/application/components/cells/DeploymentName.vue
