+++ frontend/src/pages/application/components/cells/DeviceLink.vue
