+++ frontend/src/pages/application/components/cells/InstanceInstancesLink.vue
