+++ frontend/src/pages/application/components/cells/LastSeen.vue
