+++ frontend/src/pages/application/components/cells/ProjectEditorLink.vue
