+++ frontend/src/pages/application/components/cells/Snapshot.vue
@@ -2,8 +2,8 @@
     <span class="flex space-x-4">
         <span
             v-if="activeSnapshot?.id || updateNeeded"
-            class="flex items-center space-x-2 text-gray-500 italic"
             v-ff-tooltip:left="'Target snapshot has not yet been deployed to this device.'"
+            class="flex items-center space-x-2 text-gray-500 italic"
         >
             <ExclamationIcon
                 v-if="updateNeeded"
