+++ frontend/src/pages/application/routes.js
@@ -1,56 +1,56 @@
 /**
- * WARNING: There is ongoing work to move project functionality up into applications
+ * WARNING: There is ongoing work to move Application functionality up into applications
  * or down into instances.
  *
  * No new functionality should be added here.
  */
-import Project from '@/pages/project/index.vue'
-import ProjectOverview from '@/pages/project/Overview.vue'
-import ProjectActivity from '@/pages/project/Activity.vue'
-import ProjectLogs from '@/pages/project/Logs.vue'
-import ProjectSettings from '@/pages/project/Settings.vue'
+import ApplicationActivity from '@/pages/application/Activity.vue'
+import ApplicationLogs from '@/pages/application/Logs.vue'
+import ApplicationOverview from '@/pages/application/Overview.vue'
+import ApplicationSettings from '@/pages/application/Settings.vue'
+import ApplicationIndex from '@/pages/application/index.vue'
 
 export default [
     {
-        path: '/project/:id',
+        path: '/Application/:id',
         redirect: to => {
-            return `/project/${to.params.id}/instances`
+            return `/Application/${to.params.id}/instances`
         },
-        name: 'Project',
-        component: Project,
+        name: 'Application',
+        component: ApplicationIndex,
         meta: {
             title: 'Application - Overview'
         },
         children: [
             {
                 path: 'instances',
-                component: ProjectOverview,
+                component: ApplicationOverview,
                 meta: {
                     title: 'Application - Instances'
                 }
             },
             {
                 path: 'settings',
-                component: ProjectSettings,
+                component: ApplicationSettings,
                 meta: {
                     title: 'Application - Settings'
                 }
             },
             {
                 path: 'logs',
-                component: ProjectLogs,
+                component: ApplicationLogs,
                 meta: {
                     title: 'Application - Logs'
                 }
             },
             {
                 path: 'activity',
-                component: ProjectActivity,
+                component: ApplicationActivity,
                 meta: {
                     title: 'Application - Activity'
                 }
             }
-            // { path: 'debug', component: ProjectDebug }
+            // { path: 'debug', component: ApplicationDebug }
         ]
     }
 ]
