+++ frontend/src/routes.js
@@ -5,7 +5,7 @@ import PageNotFound from '@/pages/PageNotFound'
 
 import AdminRoutes from '@/pages/admin/routes.js'
 import DeviceRoutes from '@/pages/device/routes.js'
-import ProjectRoutes from '@/pages/project/routes.js'
+import ProjectRoutes from '@/pages/application/routes.js'
 import InstanceRoutes from '@/pages/instance/routes.js'
 import TeamRoutes from '@/pages/team/routes.js'
 import AccountRoutes from '@/pages/account/routes.js'
